Initial commit
This is a GStreamer plugin that can use the color conversion and
scaling with the hardware acceleration on R-Car SoC. This feature is
available through the dmabuf method or userptr method of the V4L2 driver.
Both of the methods are used for the buffer passing between
the GStreamer plugin and the V4L2 in order to do zero-copy processing.