Explicitly specify bytesused on all OUTPUT buffers
In the latest linux kernels, specifiying 0 for the bytesused field in
vl42_buffer structs is deprecated and causes warning messages in the log.
Explicitly specify the bytesused values for all OUTPUT buffers types. This
also means that the length field buffers must be specified because the
kernel will check that bytesused <= length when OUTPUT buffers are queued.