Set strides in bytes for the planes number of GstVideoMeta
The number of GstMemory objects indicates how many separate memory
blocks a GstBuffer has, so it could be different from the planes
number corresponding to the GstVideoMeta that has multiple planes
in a contiguous memory area.
This change converts strides in GstVideoMeta to byte unit for
the planes number in the same metadata.