Add vspfilter's bufferpool
This bufferpool is created from V4L2 MMAP memory and considers that
both the output and capture interfaces use it.
Generally the MMAP memory is mapped to virtual memory space, but
this pool always exports dmabuf fds and pool's users can map them
as needed or use them without mapping.