Add a new support for the MMAP io-mode
This support makes V4L2 MMAP memory available for this plugin's
input and output.

Which the dmabuf or mmap io-mode should be used is automatically
determined by the plugin, but only the userptr io-mode needs to be
explicitly selected with the input or output io-mode property
because it can't be distinguished from the mmap memory.