Implement measure for odd size yuv buffer input.
When odd size yuv buffer come,
result of S_FMT is different from SUBDEV_S_FMT,
so STREAM_ON failed.

To make sure that both results are the same,
change the input odd size information
to round up.

Also, do cropping so that the margin of
the round up buffer is not displayed.
When caps of output size is not specified,
it will be changed to cropped size.