Add option for setting input buffer color range.
Basically, color range information is acquired from caps.
However if the upstream element does not set correctly,
it is necessary to set input color range separately.

This option works only when using external buffer for input.
This is due to implementation problems that the flag
can not be passed to function that makes pool.

Also, since full range conversion of bt709 is not supported,
When this setting is made, it fallbacks to the full range of bt601.