vspfilter: Don't store the number of buffer planes in device struct
The number of buffer planes is a property of the buffers, not the
device they are committed to.  It is also easily determined by the
video info / video meta that is already saved elsewhere. Remove the
redundant storage and refactor to use the plane count stored in the
video info.