Add 'disable-passthrough' property
By default, the vspfilter will not touch any video buffers if the
caps before and after the element are the same.  This means that
tranformations, such as input frme cropping that are not represented
in the caps will not be performed unless another transformation
(i.e color conversion or resize) is also performed.

Add a 'disable-passthrough' flag that will force processing of video
buffers regardless of the negotiated caps.