Move input window size query to separate function
For now just collect this into a basic helper function to avoid
open coding the ioctl call alongside other higher level code.

If more fields that just the width and height are needed, it would
probably be best to modify this to fill all of the format information
into something like a VideoInfo struct.