gstvspfilter: Use plane mapping frame count when using mapped frames
The number of planes in a frame will be determined by the color
format, so the same value can be obtained using both functions, but
it is best to be consistent w.r.t. which method is used.

Also add in some comments explaining why we can use macros that refer
to frame components(ie. GST_VIDEO_FRAME_COMP_*) instead of planes
when determining plane sizes.  Basically there are no macros that
take plane number as in input, but the results will be the same
whether plane number or component number is used as long as the
plane number <= comonent number (which it is).